+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 OpenStudio Analysis Gem Change Log
 ==================================
 
-Version 0.4.2 (Unreleased)
+Version 0.4.2
 -------------
 * Bug fix when adding measure from path, this now sets the correct argument name.
 * Fix namespace conflict with OpenStudio::Logger and OpenStudio::Time